chore/PRSD-NONE-rename-RegisteredPropertyDataModel-to-RegisteredPropertyViewModel
